temp

+++ internal/data/tokens.go
@@ -54,16 +54,15 @@ type TokenModel struct {
 }
 
 func (m TokenModel) Insert(token *Token) error {
-	query := `
-		INSERT INTO tokens (hash, user_id, expiry, scope)
-		VALUES ($1, $2, $3, $4)
-		`
-	args := []interface{}{token.Hash, token.UserID, token.Expiry, token.Scope}
+	query := `INSERT INTO tokens (hash, user_id, expiry, scope) VALUES ($1, $2, $3, $4)`
+
+	args := []any{token.Hash, token.UserID, token.Expiry, token.Scope}
+
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
+
 	_, err := m.DB.ExecContext(ctx, query, args...)
 	return err
-
 }
 
 func (m TokenModel) New(userID int64, ttl time.Duration, scope string) (*Token, error) {
@@ -71,16 +70,21 @@ func (m TokenModel) New(userID int64, ttl time.Duration, scope string) (*Token, 
 	if err != nil {
 		return nil, err
 	}
-	err = m.Insert(token)
-	return token, err
+
+	if err := m.Insert(token); err != nil {
+		return nil, err
+	}
+
+	return token, nil
 }
 
 func (m TokenModel) DeleteAllForUser(scope string, userID int64) error {
-	query := `
-DELETE FROM tokens
-WHERE scope = $1 AND user_id = $2`
+	query := `DELETE FROM tokens WHERE scope = $1 AND user_id = $2`
+
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
-	_, err := m.DB.ExecContext(ctx, query, scope, userID)
+
+	args := []any{scope, userID}
+	_, err := m.DB.ExecContext(ctx, query, args...)
 	return err
 }
